requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
scikit-learn==1.0.2
numpy==1.21.6
pandas==2.2.0
openpyxl==3.1.3
matplotlib==3.7.3
pyserial
imbalanced-learn